DROID-3429 ui fixes
